feat(worldgen): extract town/window helpers to new file and wire imports in town_gen.js (buildOutdoorMask, repairBuildingPerimeters, placeWindowsOnAll)
